moved the validation test to pulsarclienttests
